pyproject.toml
[project]
name = "research-assistant"
version = "1.0.0"
description = "AI research assistant built using google adk"
authors = [
    {name = "Nikunj Desai (desai.nikunjm@gmail.com)"}
]
readme = "README.md"
requires-python = "y"
dependencies = [
    "google-adk (>=1.19.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
